refactor filters module
